+++ src/StreamTransactionPkg.vhd
@@ -25,6 +25,10 @@
 --
 --  Revision History:
 --    Date      Version    Description
+--    09/2023   2023.09    Added ModelParametersIDType to Record, 
+--                         Added SendAndGet and SendAndGetBurst,
+--                         Added OperationType ENUMs:  EXTEND_DIRECTIVE_OP, EXTEND_OP, EXTEND_TX_OP, EXTEND_RX_OP
+--                         Added ClassifyUnimplementedOperation, ClassifyUnimplementedTransmitterOperation, ClassifyUnimplementedReceiverOperation
 --    05/2023   2023.05    Added SetDelayCoverageID and GetDelayCoverageID
 --    11/2022   2022.11    Added StreamRecArrayType
 --    01/2022   2022.01    Burst patterns - Burst, BurstInc, BurstRandom
@@ -38,7 +42,7 @@
 --
 --  This file is part of OSVVM.
 --  
---  Copyright (c) 2019 - 2022 by SynthWorks Design Inc.  
+--  Copyright (c) 2019 - 2023 by SynthWorks Design Inc.  
 --  
 --  Licensed under the Apache License, Version 2.0 (the "License");
 --  you may not use this file except in compliance with the License.
@@ -64,6 +68,7 @@ library osvvm ;
   context osvvm.OsvvmContext ;  
   use osvvm.ScoreboardPkg_slv.all ; 
   
+  use work.ModelParametersSingletonPkg.all ; 
   use work.FifoFillPkg_slv.all ; 
 
 package StreamTransactionPkg is 
@@ -74,7 +79,7 @@ package StreamTransactionPkg is
   --  to the model via the transaction interface
   -- ========================================================
   type StreamUnresolvedOperationType is (
-    -- Default. Used for Multiple Driver Detection
+    -- Default. Used by resolution function for Multiple Driver Detection
     NOT_DRIVEN,  
     -- Directives
     WAIT_FOR_CLOCK, 
@@ -94,12 +99,19 @@ package StreamTransactionPkg is
     -- Model Options
     SET_MODEL_OPTIONS,
     GET_MODEL_OPTIONS,
+    -- VC Customization of Directives and Functional Operations
+    EXTEND_DIRECTIVE_OP,
+    EXTEND_OP,
     --  Transmitter
+    START_OF_TX_OPS,
     SEND, 
     SEND_ASYNC,
     SEND_BURST,
     SEND_BURST_ASYNC,
+    -- VC Customization of TX Operations
+    EXTEND_TX_OP,
     -- Receiver
+    START_OF_RX_OPS,
     GET,             
     TRY_GET,
     GET_BURST,
@@ -108,7 +120,14 @@ package StreamTransactionPkg is
     TRY_CHECK,
     CHECK_BURST,
     TRY_CHECK_BURST,
-    MULTIPLE_DRIVER_DETECT  -- value used when multiple drivers are present
+    -- VC Customization of RX Operations
+    EXTEND_RX_OP,
+    -- Send and Get
+    SEND_AND_GET, 
+    SEND_AND_GET_BURST, 
+
+    -- Resolution function detected Multiple drivers
+    MULTIPLE_DRIVER_DETECT 
   ) ;
   type StreamUnresolvedOperationVectorType is array (natural range <>) of StreamUnresolvedOperationType ;
   -- Maximum is implicitly defined for any array type in VHDL-2008.   
@@ -146,6 +165,10 @@ package StreamTransactionPkg is
     ParamFromModel  : std_logic_vector_max_c ; 
     -- BurstFifo
     BurstFifo       : ScoreboardIdType ; 
+--    UseCheckFifo    : boolean_max ; 
+--    CheckFifo       : ScoreboardIdType ; 
+    -- Parameters - internal settings for the VC in a singleton data structure   
+    Params          : ModelParametersIDType ;  
     -- Verification Component Options Parameters - used by SetModelOptions
     IntToModel      : integer_max ;
     IntFromModel    : integer_max ; 
@@ -1165,6 +1188,49 @@ package StreamTransactionPkg is
     constant  StatusMsgOn      : in    boolean := false
   ) ;  
 
+  -- ========================================================
+  --  Send And Get Transactions
+  -- 
+  -- ========================================================
+  ------------------------------------------------------------
+  procedure SendAndGet (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iData            : in    std_logic_vector ;
+    constant  iParam           : in    std_logic_vector ;
+    variable  oData            : out   std_logic_vector ;
+    variable  oParam           : out   std_logic_vector ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) ;  
+
+  ------------------------------------------------------------
+  procedure SendAndGet (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iData            : in    std_logic_vector ;
+    variable  oData            : out   std_logic_vector ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) ;  
+
+  ------------------------------------------------------------
+  procedure SendAndGetBurst (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iNumFifoWords    : in    integer ;
+    constant  iParam           : in    std_logic_vector ;
+    variable  oNumFifoWords    : out   integer ;
+    variable  oParam           : out   std_logic_vector ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) ; 
+
+  ------------------------------------------------------------
+  procedure SendAndGetBurst (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iNumFifoWords    : in    integer ;
+    variable  oNumFifoWords    : out   integer ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) ; 
 
   -- ========================================================
   --  Pseudo Transactions
@@ -1208,6 +1274,39 @@ package StreamTransactionPkg is
   -----------------------------------------------------------
     constant  Operation        : in StreamOperationType
   ) return boolean ;
+
+  ------------------------------------------------------------
+  function IsTransmitterOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType
+  ) return boolean  ;
+
+  ------------------------------------------------------------
+  function IsReceiverOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType
+  ) return boolean ;
+
+  ------------------------------------------------------------
+  function ClassifyUnimplementedOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType;
+    constant  TransactionCount : in natural
+  ) return string ;
+
+  ------------------------------------------------------------
+  function ClassifyUnimplementedTransmitterOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType;
+    constant  TransactionCount : in natural
+  ) return string ;
+  
+  ------------------------------------------------------------
+  function ClassifyUnimplementedReceiverOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType ;
+    constant  TransactionCount : in natural
+  ) return string ;
 
 end StreamTransactionPkg ;
 
@@ -2621,6 +2720,70 @@ package body StreamTransactionPkg is
     TryCheckBurstRandom(TransactionRec, CoverID, NumFifoWords, FifoWidth, "", Available, StatusMsgOn) ;
   end procedure TryCheckBurstRandom ;  
 
+
+  -- ========================================================
+  --  Send And Get Transactions
+  -- 
+  -- ========================================================
+  ------------------------------------------------------------
+  procedure SendAndGet (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iData            : in    std_logic_vector ;
+    constant  iParam           : in    std_logic_vector ;
+    variable  oData            : out   std_logic_vector ;
+    variable  oParam           : out   std_logic_vector ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) is 
+  begin
+    LocalSend(TransactionRec, SEND_AND_GET, iData, iParam, StatusMsgOn) ;
+    oData  := SafeResize(TransactionRec.DataFromModel,  oData'length) ; 
+    oParam := SafeResize(TransactionRec.ParamFromModel, oParam'length) ; 
+  end procedure SendAndGet ;  
+
+  ------------------------------------------------------------
+  procedure SendAndGet (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iData            : in    std_logic_vector ;
+    variable  oData            : out   std_logic_vector ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) is 
+  begin
+    LocalSend(TransactionRec, SEND_AND_GET, iData, "", StatusMsgOn) ;
+    oData  := SafeResize(TransactionRec.DataFromModel, oData'length) ; 
+  end procedure SendAndGet ;  
+
+  ------------------------------------------------------------
+  procedure SendAndGetBurst (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iNumFifoWords    : in    integer ;
+    constant  iParam           : in    std_logic_vector ;
+    variable  oNumFifoWords    : out   integer ;
+    variable  oParam           : out   std_logic_vector ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) is 
+  begin
+    LocalSendBurst(TransactionRec, SEND_AND_GET_BURST, iNumFifoWords, iParam, StatusMsgOn) ;
+    oNumFifoWords := TransactionRec.IntFromModel ;
+    oParam        := SafeResize(TransactionRec.ParamFromModel, oParam'length) ; 
+  end procedure SendAndGetBurst ; 
+
+  ------------------------------------------------------------
+  procedure SendAndGetBurst (
+  ------------------------------------------------------------
+    signal    TransactionRec   : inout StreamRecType ;
+    constant  iNumFifoWords    : in    integer ;
+    variable  oNumFifoWords    : out   integer ;
+    constant  StatusMsgOn      : in    boolean := false 
+  ) is 
+  begin
+    LocalSendBurst(TransactionRec, SEND_AND_GET_BURST, iNumFifoWords, "", StatusMsgOn) ;
+    oNumFifoWords := TransactionRec.IntFromModel ;
+  end procedure SendAndGetBurst ; 
+
+
   -- ========================================================
   --  Pseudo Transactions
   --  Interact with the record only.
@@ -2687,5 +2850,77 @@ package body StreamTransactionPkg is
   begin
     return (Operation = CHECK) or (Operation = TRY_CHECK) or (Operation = CHECK_BURST) or (Operation = TRY_CHECK_BURST) ;
   end function IsCheck ;
+  
+  ------------------------------------------------------------
+  function IsTransmitterOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType
+  ) return boolean is
+  begin
+    return (Operation < START_OF_RX_OPS) ;
+  end function IsTransmitterOperation ;
+
+  ------------------------------------------------------------
+  function IsReceiverOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType
+  ) return boolean is
+  begin
+    return (Operation < START_OF_TX_OPS) or (Operation > START_OF_RX_OPS) ;
+  end function IsReceiverOperation ;
+
+  ------------------------------------------------------------
+  function ClassifyUnimplementedOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType;
+    constant  TransactionCount : in natural
+  ) return string is
+  begin
+    if Operation = MULTIPLE_DRIVER_DETECT then
+      return "Multiple Drivers on Transaction Record." & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    else
+      return "Unimplemented Transaction: " & to_string(Operation) & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    end if ; 
+  end function ClassifyUnimplementedOperation ;
+
+  ------------------------------------------------------------
+  function ClassifyUnimplementedTransmitterOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType;
+    constant  TransactionCount : in natural
+  ) return string is
+  begin
+    if Operation = MULTIPLE_DRIVER_DETECT then
+      return "Multiple Drivers on Transaction Record." & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    elsif IsReceiverOperation(Operation) then
+      return "Not a Transmitter Transaction: " & to_string(Operation) & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    else
+      return "Unimplemented Transaction: " & to_string(Operation) & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    end if ; 
+  end function ClassifyUnimplementedTransmitterOperation ;
+  
+  ------------------------------------------------------------
+  function ClassifyUnimplementedReceiverOperation (
+  -----------------------------------------------------------
+    constant  Operation        : in StreamOperationType ;
+    constant  TransactionCount : in natural
+  ) return string is
+  begin
+    if Operation = MULTIPLE_DRIVER_DETECT then
+      return "Multiple Drivers on Transaction Record." & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    elsif IsTransmitterOperation(Operation) then
+      return "Not a Receiver Transaction: " & to_string(Operation) & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    else
+      return "Unimplemented Transaction: " & to_string(Operation) & 
+             "  Transaction # " & to_string(TransactionCount) ;
+    end if ; 
+  end function ClassifyUnimplementedReceiverOperation ;
 
 end StreamTransactionPkg ;